Fix wrong response dispatching on wrapped rq_seq_nr
This is most likely to happen when automatic SEL polling is enabled.

If a packet gets lost/discarded the receiver could in rare cases receive
the answer of an unrelated request. This is due a match on the request
sequence number only which could have wrapped and reused if many
concurrent requests are in progress.